Modify: user, order, product models.py 수정 및 review app 삭제
