+++ product/models.py
@@ -1,10 +1,12 @@
 from django.db import models
 
+
 class Category(models.Model):
     name = models.CharField(max_length=45)
 
     class Meta:
         db_table = 'categories'
+
 
 class SubCategory(models.Model):
     name     = models.CharField(max_length=30)
@@ -13,27 +15,29 @@ class SubCategory(models.Model):
     class Meta:
         db_table = 'sub_categories'
 
+
 class Product(models.Model):
     sub_category    = models.ForeignKey('SubCategory', on_delete=models.PROTECT)
     name            = models.CharField(max_length=45)
     description     = models.CharField(max_length=60)
-    price           = models.IntegerField()
+    price           = models.DecimalField(max_digits=10, decimal_places=2)
     image_url       = models.CharField(max_length=3000)
     sales_unit      = models.CharField(max_length=30, null=True)
-    amount          = models.CharField(max_length=30, null=True) # 중량/용량
+    amount          = models.CharField(max_length=30, null=True) 
     origin          = models.CharField(max_length=30, null=True)
-    storage_method  = models.ForeignKey('StorageMethod', on_delete=models.SET_NULL, null=True) #DO_NOTHING ? 
-    expiration_date = models.CharField(max_length=100, null=True) # 정해진 날짜가 없고 대부분 짧은 텍스트로 설명이 되어있음
-    stock           = models.IntegerField()
+    storage_method  = models.ForeignKey('StorageMethod', on_delete=models.SET_NULL, null=True)  
+    expiration_date = models.CharField(max_length=100, null=True) 
+    stock           = models.PositiveIntegerField(default=0)
     content         = models.TextField()
     uploaded_at     = models.DateField(auto_now_add=True)
     product_options = models.ManyToManyField('self',
                                             through='ProductOption',
                                             symmetrical=False,
-                                            related_name='options') # related_name의 필요성?, symmetrical true, false의 정확한 사용, self ManyToMany
+                                            related_name='options') 
     
     class Meta:
         db_table = 'products'
+
 
 class ProductOption(models.Model):
     from_product_id = models.ForeignKey('Product', on_delete=models.CASCADE, related_name='from_product')
@@ -42,20 +46,23 @@ class ProductOption(models.Model):
     class Meta:
         db_table = 'product_options'
 
+
 class DiscountRate(models.Model):
     product       = models.ForeignKey('Product', on_delete=models.CASCADE)
-    discount_rate = models.IntegerField()
+    discount_rate = models.DecimalField(max_digits=10, decimal_places=5)
 
     class Meta:
         db_table = 'discount_rate'
 
+
 class DailySpecialDiscount(models.Model):
     product             = models.ForeignKey('Product', on_delete=models.CASCADE)
     daily_discount_rate = models.IntegerField()
-    start_date          = models.DateTimeField() # DateTimeField or DateField ? time을 같이 넘겨주거나, time은 프론트에서 처리하거나 더 나은 선택은?
+    start_date          = models.DateField()
 
     class Meta:
         db_table = 'daily_special_discount'
+
 
 class Allergy(models.Model):
     name = models.CharField(max_length=45)
@@ -64,6 +71,7 @@ class Allergy(models.Model):
     class Meta:
         db_table = 'allergies'
 
+
 class StorageMethod(models.Model):
     name = models.CharField(max_length=8)
 
@@ -71,28 +79,23 @@ class StorageMethod(models.Model):
         db_table = 'storage_methods'
 
 
+class Review(models.Model):
+    user       = models.ForeignKey('user.User', on_delete=models.CASCADE)
+    product    = models.ForeignKey('Product', on_delete=models.CASCADE)
+    order      = models.ForeignKey('order.Order', on_delete=models.CASCADE) 
+    image_url  = models.CharField(max_length=2000, null=True)
+    content    = models.CharField(max_length=100)
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True, null=True)
+
+    class Meta:
+        db_table = 'reviews'
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+class ReviewLike(models.Model):
+    user     = models.ForeignKey('user.User', on_delete=models.CASCADE)
+    review   = models.ForeignKey('Review', on_delete=models.CASCADE)
+    is_liked = models.BooleanField(default=False)
+    
+    class Meta:
+        db_table = 'review_likes'
